[SPR-85] refactor: JwtTokenProvider 위치 변경
